requirements.txt
Pillow>=9.0.0
qrcode[pil]>=7.0.0 